Update PackageProjectUrl in Kimi.NetExtensions.csproj
The URL for the `PackageProjectUrl` property in the `Kimi.NetExtensions.csproj` file was updated. The old URL `https://github.com/kamazheng/Kimi.MudBlazorExtentions` was replaced with the new URL `https://github.com/kamazheng/Kimi.NetExtensions`.

This change ensures that the project file points to the correct GitHub repository for the `Kimi.NetExtensions` project, correcting the previously incorrect or outdated URL.